LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/back-forward-opaque-origin-expected.txt

CONSOLE MESSAGE: Origin null is not allowed by Access-Control-Allow-Origin. Status code: 200


Harness Error (TIMEOUT), message = null
